backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5

python-magic>=0.4.27
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23

weasyprint>=60.0

httpx>=0.26.0

pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

python-dotenv>=1.0.0

opencode>=0.0.0
